1_Models/1_LanguageModels/1.2_ChatModels/1.2.1_ClosedSourceModels/1.2.1.3_Google/Google.py
from langchain_google_genai import ChatGoogleGenerativeAI
from dotenv import load_dotenv
load_dotenv()

model = ChatGoogleGenerativeAI(model="gemini-3-flash-preview", temperature=0.7)
response = model.invoke("What is the difference between and purpose of langchain, langgraph, langsmith and deepagent?")
print(response.content)