pyproject.toml
[project]
name = "1-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-generativeai>=0.8.6",
    "huggingface-hub>=1.8.0",
    "langchain>=1.2.14",
    "langchain-anthropic>=1.4.0",
    "langchain-core>=1.2.23",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "numpy>=2.4.4",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "torch>=2.11.0",
    "transformers>=5.4.0",
]
